Initial server setup.
The first commit contains the general setup for the server side of the
application. This includes the config files for TypeScript, TypeORM, and
the package file for the project. Also have a basic setup for the server
including UserController which will ultimately be changed to a class
rather than separate route listeners so each route can be implemented in
a method.